+++ eras/shelley/impl/CHANGELOG.md
@@ -3,7 +3,11 @@
 ## 1.9.0.0
 
 * Change the type of `ppEMaxL`
-	
+
+
+### `testlib`
+
+* Add `runImpTestM`, `runImpTestM_`, `evalImpTestM` and `execImpTestM`
 
 ## 1.8.0.0